chore: move hook to hooks dir
