+++ application/ui/src/features/annotator/hooks/use-is-annotator-scene-busy.ts
@@ -1,7 +1,7 @@
 // Copyright (C) 2025-2026 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { useVideoPlayerContext } from './video-player/video-player-provider.component';
+import { useVideoPlayerContext } from '../video-player/video-player-provider.component';
 
 export const useIsAnnotatorSceneBusy = () => {
     const context = useVideoPlayerContext();
